feat: move template inside story
