+++ src/stories/columns-caption-with-styles.stories.tsx
@@ -1,11 +1,46 @@
-import { createBaseStory, Template } from './base';
-import { createDateInstance } from '../index';
+import React from 'react';
+import { createBaseStory, BaseTemplate } from './base';
+import {
+  createDateInstance,
+  format,
+  Picker,
+  WheelPickerSelectEvent,
+} from '../index';
 // Types
 import type { DatePickerConfig } from '../index';
+import { ComponentStory } from '@storybook/react';
+import { PickerProps } from '../index.types';
+import { action } from '@storybook/addon-actions';
+import type { Event } from '../components/WheelPicker/index.types';
 
 export default createBaseStory('Columns Caption With Styles');
 
+const Template: ComponentStory<typeof Picker> = (args) => {
+  const pickerProps = { theme: 'light', ...args } as Required<PickerProps>;
+  const [selectedDateValue, setSelectedDateValue] = React.useState<string>();
+  const [selectedDateEvents, setSelectedDateEvents] = React.useState<
+    Array<Event>
+  >([]);
+
+  function handleOnChange(data: WheelPickerSelectEvent) {
+    setSelectedDateValue(format(data.date!, 'd MMMM yyyy'));
+    setSelectedDateEvents(data.events);
+    action('onClick')(data);
+  }
+
+  return (
+    <BaseTemplate value={selectedDateValue!} events={selectedDateEvents!}>
+      <Picker
+        {...pickerProps}
+        onChange={handleOnChange}
+        onSubmit={handleOnChange}
+      />
+    </BaseTemplate>
+  );
+};
+
 export const ColumnsCaptionWithStyles = Template.bind({});
+
 const ColumnsCaptionWithStyleConfig: DatePickerConfig = {
   year: {
     caption: {
